fix: Resolve sign comparison warnings
Cast long values to size_t when comparing with fread/fwrite return values
and GetHandleSize results:

- tcpverbs.c:676: Cast bytes_to_read to size_t for GetHandleSize comparison
- fileverbs_portable.c:1537: Cast filesize to size_t for bytesread comparison
- fileverbs_portable.c:1580: Cast datasize to size_t for fwrite comparison

All values have already been validated as positive before comparison,
making the unsigned cast safe.

+++ Common/source/tcpverbs.c
@@ -673,7 +673,7 @@ boolean tcp_read_stream(long stream_id, long bytes_to_read, Handle *data_out) {
     }
 
     /* SECURITY: Verify allocated buffer size matches request (prevents undersized allocation) */
-    if (GetHandleSize(hdata) < bytes_to_read) {
+    if (GetHandleSize(hdata) < (size_t) bytes_to_read) {
         disposehandle(hdata);
         tcp_stream_release(stream);
         *data_out = nil;
